beginner/submissions/team-members/art-turner/pyproject.toml
[project]
name = "scholarai-beginner"
version = "0.1.0"
description = "ScholarAI Beginner Track - AI-powered research assistant"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools", "models", "exporters"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
